fix: files cmd
